HtmlBridge Phase 3 ratchet: de-list Utilities.cs (split name-validation partial)
Peel the cohesive DOM name-validation cluster and the JS-side constructor
globals it validates against out of DomBridge/Utilities.cs into a sibling
partial DomBridge/Utilities.NameValidation.cs (286 lines): ThrowDOMException,
ValidateElementName / ValidateQualifiedName, RegisterDOMException,
RegisterNodeConstructor, RegisterSVGLength, and the two [GeneratedRegex]
XML-name patterns with their backing fields.

Pure partial-class relocation — no signature, accessibility, or logic change,
so behaviour-identical by construction. Utilities.cs drops 894 -> 626 lines,
under the 750-line guard, so its OversizedFileExemptions entry is removed and
the stale-exemption check now enforces it stays under the limit. Debt list:
eight -> seven files.

Broiler.HtmlBridge.Dom builds clean; HtmlBridgeArchitectureGuardTests 14/14
green (0 offenders / 0 stale exemptions); NamespaceAndDomCoreTests /
DomImplementationTests name-validation coverage (27 tests) green.
